Add support for OrderBook10 WS stream
The type of the data field for the WS streams
is equivalent to the corresponding REST endpoint type.
However, some of them (like OrderBook10) don't have an
equivalent and have to be implemented. Actually, it might
be a good idea to re-implement the existing REST endpoint
types, for both consistency and quality reasons. The
types provided by the auto-generated REST API are complete shit.